Task2

+++ Tyuiu.StepanyukMV.Sprint1.Task2.V25.Test/DataServiceTest.cs
@@ -8,7 +8,7 @@ namespace Tyuiu.StepanyukMV.Sprint1.Task2.V25.Test
         public void ValidExpression()
         {
             DataService ds = new DataService();
-            int x = 2;
+            double x = 2;
             var res = ds.ConvertRadsToDegrees(x);
             Assert.AreEqual(%3, res);
         }
